requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.2
python-dotenv>=1.0.0
pydantic>=2.11.4
PyYAML>=6.0.1
livekit-agents>=0.3.0
livekit-plugins-openai>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-cartesia>=1.0.0
livekit-plugins-deepgram>=1.0.0
aiohttp>=3.9.3 